make it the price and amount float64

+++ internal/database/menu.go
@@ -4,11 +4,11 @@ type Menu struct {
 	Id          int     `json:"id"`
 	Name        string  `json:"name"`
 	Description string  `json:"description"`
-	Price       float32 `json:"price"`
+	Price       float64 `json:"price"`
 	//category
 }
 
-func NewMenu(id int, name string, description string, price float32) Menu {
+func NewMenu(id int, name string, description string, price float64) Menu {
 	return Menu{
 		Id:          id,
 		Name:        name,
